tuning: move schema to lenskit.schemas
